requirements.txt
python-dotenv
watchdog
python-docx
plyer
winotify
